[Refactor] 버킷 생성 템플릿 스타일 파일 리
