+++ sidebar/src/components/templates/bucket_create_template/index.js
@@ -1,25 +1,8 @@
 import React from 'react';
-import styled from 'styled-components';
-import { makeStyles } from '@material-ui/core/styles';
 import BucketContents from '../../UI/organisms/bucket_contents';
 import CreateDetailList from '../../UI/organisms/create_detail_list';
 import Buttons from '../../UI/organisms/buttons';
-
-const BucketCreateTemplateWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  padding: 15px;
-`;
-
-const useStyles = makeStyles((theme) => ({
-  header: {
-    display: 'flex',
-    alignItems: 'center',
-    padding: theme.spacing(0, 1),
-    ...theme.mixins.toolbar,
-    justifyContent: 'flex-end',
-  },
-}));
+import useStyles, { BucketCreateTemplateWrapper } from './style';
 
 const BucketCreateTemplate = () => {
   const classes = useStyles();
